requirements/base_requirements.txt
attrs==19.3.0
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.2.5
django-allauth==0.40.0
django-braces==1.13.0
django-ckeditor==5.8.0
django-cleanup==4.0.0
django-embed-video==1.3
django-environ==0.4.5
django-jet==1.0.8
django-js-asset==1.2.2
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-cors-headers==3.2.1
django-rest-passwordreset==1.0.0
django-rest-swagger==2.2.0
django-taggit==1.1.0
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
drf-yasg==1.17.0
google-api-python-client==1.4.1
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==19.2
pyparsing==2.4.5
pyrsistent==0.15.6
python3-openid==3.1.0
pytz==2019.2
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.5
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
swagger-spec-validator==2.4.3
uritemplate==3.0.0
urllib3==1.25.7
django-taggit-serializer==0.1.7